+++ src/Common/Constants.ts
@@ -12,6 +12,7 @@ export default class Constants {
   public static readonly defaultChannelId = ChannelId.LCW;
   public static readonly defaultLocale = "en-us";
   public static readonly noContentStatusCode = 204;
+  public static readonly tooManyRequestsStatusCode = 429;
   public static readonly sensitiveProperties = ["AuthenticatedUserToken"];
   public static readonly transactionid = "transaction-id";
   public static readonly customerDisplayName = "customerDisplayName";